销售端登录时添加用户角色判断

+++ dangjia-group/dangjia-master/dangjia-service-master/src/main/java/com/dangjia/acg/service/member/MemberService.java
@@ -19,12 +19,16 @@ import com.dangjia.acg.mapper.core.IWorkerTypeMapper;
 import com.dangjia.acg.mapper.house.IHouseDistributionMapper;
 import com.dangjia.acg.mapper.house.IHouseMapper;
 import com.dangjia.acg.mapper.member.*;
+import com.dangjia.acg.mapper.store.IStoreMapper;
+import com.dangjia.acg.mapper.store.IStoreUserMapper;
 import com.dangjia.acg.mapper.user.UserMapper;
 import com.dangjia.acg.modle.config.Sms;
 import com.dangjia.acg.modle.core.WorkerType;
 import com.dangjia.acg.modle.house.House;
 import com.dangjia.acg.modle.house.HouseDistribution;
 import com.dangjia.acg.modle.member.*;
+import com.dangjia.acg.modle.store.Store;
+import com.dangjia.acg.modle.store.StoreUser;
 import com.dangjia.acg.modle.sup.Supplier;
 import com.dangjia.acg.modle.user.MainUser;
 import com.dangjia.acg.service.activity.RedPackPayService;
@@ -98,14 +102,16 @@ public class MemberService {
     private RedisClient redisClient;
     @Autowired
     private MessageAPI messageAPI;
+    @Autowired
+    private IStoreMapper iStoreMapper;
+    @Autowired
+    private IStoreUserMapper iStoreUserMapper;
 
     /**
      * 获取用户手机号
      *
-     * @param request
-     * @param id      来源ID
-     * @param idType  1=房屋ID, 2=用户ID, 3=供应商ID, 4=系统用户, 5=验房分销
-     * @return
+     * @param id     来源ID
+     * @param idType 1=房屋ID, 2=用户ID, 3=供应商ID, 4=系统用户, 5=验房分销
      */
     public ServerResponse getMemberMobile(HttpServletRequest request, String id, String idType) {
         String mobile = "";
@@ -152,7 +158,7 @@ public class MemberService {
     /**
      * 获取用户详细资料
      */
-    public ServerResponse getMemberInfo(String userToken) {
+    public ServerResponse getMemberInfo(Integer userToken) {
         AccessToken accessToken = redisClient.getCache(userToken + Constants.SESSIONUSERID, AccessToken.class);
         if (accessToken == null) {//无效的token
             return ServerResponse.createbyUserTokenError();
@@ -167,7 +173,7 @@ public class MemberService {
     }
 
     // 登录 接口
-    public ServerResponse login(String phone, String password, String userRole) {
+    public ServerResponse login(String phone, String password, Integer userRole) {
         //指定角色查询用户
         Member user = new Member();
         user.setMobile(phone);
@@ -181,28 +187,59 @@ public class MemberService {
         }
     }
 
-    ServerResponse getUser(Member user, String userRole) {
-        if ("1".equals(userRole)) {
+    ServerResponse getUser(Member user, Integer userRole) {
+        if (userRole == 1) {
             clueService.sendUser(user, user.getMobile());
         }
-        MainUser mainUser = userMapper.findUserByMobile(user.getMobile());
-        if (mainUser != null) {
-            if (CommonUtil.isEmpty(mainUser.getMemberId())) {
-                //插入MemberId
-                userMapper.insertMemberId(user.getMobile());
-            }
-        } else if ("3".equals(userRole)) {
-            return ServerResponse.createByErrorMessage("当前用户暂无权限登录，请联系管理员");
+        ServerResponse serverResponse = setAccessToken(user, userRole);
+        if (!serverResponse.isSuccess()) {
+            return serverResponse;
         }
         updateOrInsertInfo(user.getId(), String.valueOf(userRole), user.getPassword());
-        userRole = "role" + userRole + ":" + user.getId();
-        String token = redisClient.getCache(userRole, String.class);
-        //如果用户存在usertoken则清除原来的token数据
-        if (!CommonUtil.isEmpty(token)) {
-            redisClient.deleteCache(token + Constants.SESSIONUSERID);
+        groupInfoService.registerJGUsers("zx", new String[]{user.getId()}, new String[1]);
+        groupInfoService.registerJGUsers("gj", new String[]{user.getId()}, new String[1]);
+        return ServerResponse.createBySuccess("登录成功，正在跳转", serverResponse.getResultObj());
+    }
+
+    private ServerResponse setAccessToken(Member user, Integer userRole) {
+        MainUser mainUser = userMapper.findUserByMobile(user.getMobile());
+        if (mainUser != null && CommonUtil.isEmpty(mainUser.getMemberId())) {
+            //插入MemberId
+            userMapper.insertMemberId(user.getMobile());
+        }
+        if (userRole == 3) {
+            if (mainUser == null) {
+                return ServerResponse.createByErrorMessage("当前用户暂无权限使用该终端，请联系管理员");
+            }
+            if (mainUser.getIsJob()) {
+                return ServerResponse.createByErrorMessage("当前用户已离职，请您联系管理员");
+            }
         }
         user.initPath(configUtil.getValue(SysConfig.PUBLIC_DANGJIA_ADDRESS, String.class));
         AccessToken accessToken = TokenUtil.generateAccessToken(user, mainUser);
+        if (userRole == 3) {
+            accessToken.setMemberType(-1);
+            Example example = new Example(Store.class);
+            example.createCriteria().andEqualTo(Store.USER_ID, mainUser.getId());
+            int c = iStoreMapper.selectCountByExample(example);
+            if (c > 0) {
+                accessToken.setMemberType(3);
+            } else {
+                example = new Example(StoreUser.class);
+                example.createCriteria().andEqualTo(StoreUser.USER_ID, mainUser.getId());
+                List<StoreUser> storeUsers = iStoreUserMapper.selectByExample(example);
+                if (storeUsers.size() > 0) {
+                    StoreUser storeUser = storeUsers.get(0);
+                    if (storeUser.getType() == 0 || storeUser.getType() == 1) {
+                        accessToken.setMemberType(storeUser.getType() == 0 ? 4 : 5);
+                    } else {
+                        return ServerResponse.createByErrorMessage("当前用户暂无权限使用该终端，请联系管理员");
+                    }
+                } else {
+                    return ServerResponse.createByErrorMessage("当前用户暂无权限使用该终端，请联系管理员");
+                }
+            }
+        }
         if (!CommonUtil.isEmpty(user.getWorkerTypeId())) {
             WorkerType wt = workerTypeMapper.selectByPrimaryKey(user.getWorkerTypeId());
             if (wt != null) {
@@ -210,10 +247,14 @@ public class MemberService {
             }
         }
         redisClient.put(accessToken.getUserToken() + Constants.SESSIONUSERID, accessToken);
-        redisClient.put(userRole, accessToken.getUserToken());
-        groupInfoService.registerJGUsers("zx", new String[]{accessToken.getMemberId()}, new String[1]);
-        groupInfoService.registerJGUsers("gj", new String[]{accessToken.getMemberId()}, new String[1]);
-        return ServerResponse.createBySuccess("登录成功，正在跳转", accessToken);
+        String userRoleText = "role" + userRole + ":" + user.getId();
+        String token = redisClient.getCache(userRoleText, String.class);
+        //如果用户存在usertoken则清除原来的token数据
+        if (!CommonUtil.isEmpty(token)) {
+            redisClient.deleteCache(token + Constants.SESSIONUSERID);
+        }
+        redisClient.put(userRoleText, accessToken.getUserToken());
+        return ServerResponse.createBySuccess(accessToken);
     }
 
     /*
@@ -268,7 +309,6 @@ public class MemberService {
             password, String invitationCode, Integer userRole) {
         Integer registerCode = redisClient.getCache(Constants.SMS_CODE + phone, Integer.class);
         if (registerCode == null || smscode != registerCode) {
-
             return ServerResponse.createByErrorMessage("验证码错误");
         } else {
             Member user = new Member();
@@ -291,22 +331,17 @@ public class MemberService {
             user.setInviteNum(0);
             user.setIsCrowned(0);
             user.setHead("qrcode/logo.png");
+            updateOrInsertInfo(user.getId(), String.valueOf(userRole), user.getPassword());
+            user.initPath(configUtil.getValue(SysConfig.PUBLIC_DANGJIA_ADDRESS, String.class));
+            ServerResponse serverResponse = setAccessToken(user, userRole);
+            if (!serverResponse.isSuccess()) {
+                return serverResponse;
+            }
             memberMapper.insertSelective(user);
 //            userRole", value = "app应用角色  1为业主角色，2为工匠角色，0为业主和工匠双重身份角色
             if (userRole == 1) {
                 clueService.sendUser(user, user.getMobile());
             }
-            updateOrInsertInfo(user.getId(), String.valueOf(userRole), user.getPassword());
-            user.initPath(configUtil.getValue(SysConfig.PUBLIC_DANGJIA_ADDRESS, String.class));
-            MainUser mainUser = userMapper.findUserByMobile(user.getMobile());
-            AccessToken accessToken = TokenUtil.generateAccessToken(user, mainUser);
-            if (!CommonUtil.isEmpty(user.getWorkerTypeId())) {
-                WorkerType wt = workerTypeMapper.selectByPrimaryKey(user.getWorkerTypeId());
-                if (wt != null) {
-                    accessToken.setWorkerTypeName(wt.getName());
-                }
-            }
-            redisClient.put(accessToken.getUserToken() + Constants.SESSIONUSERID, accessToken);
             redisClient.deleteCache(Constants.SMS_CODE + phone);
             if (userRole == 1) {
                 try {
@@ -317,7 +352,7 @@ public class MemberService {
                     logger.error("注册送优惠券活动异常-zhuce：原因：" + e.getMessage(), e);
                 }
             }
-            return ServerResponse.createBySuccess("注册成功", accessToken);
+            return ServerResponse.createBySuccess("注册成功", serverResponse.getResultObj());
         }
     }
 
@@ -343,7 +378,7 @@ public class MemberService {
     /**
      * 工匠提交详细资料
      */
-    public ServerResponse updateWokerRegister(Member user, String userToken, String userRole) {
+    public ServerResponse updateWokerRegister(Member user, String userToken) {
         HttpServletRequest request = ((ServletRequestAttributes) RequestContextHolder.getRequestAttributes())
                 .getRequest();
         AccessToken accessToken = redisClient.getCache(userToken + Constants.SESSIONUSERID, AccessToken.class);
@@ -375,7 +410,6 @@ public class MemberService {
             user.setCreateDate(null);
             memberMapper.updateByPrimaryKeySelective(user);
             user = memberMapper.selectByPrimaryKey(user.getId());
-
             user.initPath(configUtil.getValue(SysConfig.PUBLIC_DANGJIA_ADDRESS, String.class));
             MainUser mainUser = userMapper.findUserByMobile(user.getMobile());
             accessToken = TokenUtil.generateAccessToken(user, mainUser);
@@ -408,7 +442,6 @@ public class MemberService {
      * @param idcaodb   身份证反面
      * @param idcaodall 半身照
      * @param idnumber  身份证号
-     * @return
      */
     public ServerResponse certification(String userToken, String name, String idcaoda, String idcaodb, String
             idcaodall, String idnumber) {
@@ -466,7 +499,6 @@ public class MemberService {
      *
      * @param userToken    token
      * @param workerTypeId 工种类型的id
-     * @return
      */
     public ServerResponse certificationWorkerType(String userToken, String workerTypeId) {
         Object object = constructionService.getMember(userToken);
@@ -498,9 +530,6 @@ public class MemberService {
 
     /**
      * 找回密码 获取code
-     *
-     * @return
-     * @throws Exception
      */
     public ServerResponse forgotPasswordCode(String phone) {
         Member user = new Member();
@@ -524,9 +553,6 @@ public class MemberService {
 
     /**
      * 找回密码校验验证码
-     *
-     * @return
-     * @throws Exception
      */
     public ServerResponse checkForgotPasswordCode(String phone, int smscode) {
         Member user = new Member();
@@ -544,9 +570,6 @@ public class MemberService {
 
     /**
      * 找回密码更新密码
-     *
-     * @return
-     * @throws Exception
      */
     public ServerResponse updateForgotPassword(String phone, String password, String token) {
         if (CommonUtil.isEmpty(token)) {
@@ -596,13 +619,13 @@ public class MemberService {
             }
             String[] childsLabelIdArr = new String[childsLabelIdList.size()];
             childsLabelIdList.toArray(childsLabelIdArr);
-            PageHelper.startPage(pageDTO.getPageNum(), pageDTO.getPageSize());
             if (!CommonUtil.isEmpty(beginDate) && !CommonUtil.isEmpty(endDate)) {
                 if (beginDate.equals(endDate)) {
                     beginDate = beginDate + " " + "00:00:00";
                     endDate = endDate + " " + "23:59:59";
                 }
             }
+            PageHelper.startPage(pageDTO.getPageNum(), pageDTO.getPageSize());
             List<Member> list = memberMapper.getMemberListByName(searchKey, stage, userRole, childsLabelIdArr, orderBy, type, userId, beginDate, endDate);
             PageInfo pageResult = new PageInfo(list);
             List<MemberCustomerDTO> mcDTOList = new ArrayList<>();
@@ -717,8 +740,6 @@ public class MemberService {
 
     /**
      * 我的邀请码
-     *
-     * @return
      */
     public ServerResponse getMyInvitation(String userToken) {
         Object object = constructionService.getMember(userToken);
@@ -736,10 +757,8 @@ public class MemberService {
     /**
      * 查询用户实名认证列表
      *
-     * @param pageDTO
      * @param searchKey     申请人或申请人手机好关键字
      * @param realNameState -1:全部，1:认证中，2:认证被驳回，3:认证通过 （默认-1）
-     * @return
      */
     public ServerResponse certificationList(PageDTO pageDTO, String searchKey, Integer realNameState) {
         try {
